feat(mcp): add terraform mcp server (#98)
# 📥 Pull Request

## ❓ What are you trying to address

This pull request introduces a configuration file for `.vscode` to
integrate and run a Terraform MCP Server using Docker. The most
important change is the addition of the `mcp.json` file, which specifies
the command and arguments required to execute the server.

Configuration for Terraform MCP Server:

*
[`.vscode/mcp.json`](diffhunk://#diff-756c99ad6a15758981c15d6ba32a2d83391908e8e021f0399e9d34339dc4eadaR1-R14):
Added a new configuration file to define the setup for running the
Terraform MCP Server via Docker. This includes specifying the `docker`
command and its arguments to execute the
`hashicorp/terraform-mcp-server` image.